refactor: delete $ValuesType, rename ValuesType -> Values
Before, we had two identical types (`$ValuesType` and `ValuesType`),
which was pointless. So I deleted the first (since the $ types are
supposed to only mirror Flow utilities) and renamed the second, since
`Type` at the end of the name says nothing.

So, now there’s $Values, which only works on objects (as before), and
Values, which works on objects and tuples.